Make Odin Recipes Website
Create an HTML website with three recipe anchor tags that link to
recipe HTML pages containing a header, an image, a description of
the finished product, an unordered list of ingredients, and an
ordered list of recipe completion steps.

Add a 'Return to main' link at the bottom of the page.